@@ -1,6 +1,6 @@
 {
     "name": "sambal-cli",
-    "version": "0.0.33",
+    "version": "0.0.34",
     "description": "Sambal CLI tools",
     "main": "dist/src/index.js",
     "repository": {
@@ -17,7 +17,7 @@
         "static site generator"
     ],
     "engines": {
-        "node": ">=8.0.0"
+        "node": ">=10.0.0"
     },
     "bin": {
         "sambal": "./bin/sambal"
@@ -45,26 +45,30 @@
         "ws": "7.2.3"
     },
     "devDependencies": {
-        "@babel/core": "7.7.7",
-        "@babel/preset-env": "7.7.7",
+        "@babel/core": "7.9.6",
+        "@babel/preset-env": "7.9.6",
+        "@babel/preset-typescript": "7.9.0",
         "@fortawesome/fontawesome-free": "5.10.1",
         "@types/cheerio": "0.22.12",
         "@types/express": "4.17.1",
         "@types/node": "10.12.10",
+        "@types/jest": "25.2.1",
         "@types/js-yaml": "3.11.2",
         "@types/lodash": "4.14.118",
         "@types/sharp": "0.25.0",
         "@types/shelljs": "0.8.5",
         "@types/webpack": "4.41.4",
         "@types/ws": "7.2.4",
+        "babel-jest": "25.5.1",
         "babel-loader": "8.0.6",
         "bootstrap": "4.3.1",
         "css-loader": "3.4.0",
         "file-loader": "5.0.2",
-        "jquery": "3.4.1",
+        "jest": "25.5.4",
+        "jquery": "3.5.1",
         "prismjs": "1.17.1",
         "rimraf": "3.0.0",
-        "sambal": "~0.0.30",
+        "sambal": "~0.0.33",
         "style-loader": "1.1.1",
         "tslint": "5.11.0",
         "ts-node": "7.0.1",
